+++ src/lib.rs
@@ -3,6 +3,7 @@
 pub mod api;
 pub mod authelia;
 pub mod bbtw;
+pub mod builds;
 pub mod db;
 pub mod db_fields;
 pub mod dependency_graph;