+++ tests/integration/dependency_graph.rs
@@ -11,7 +11,7 @@ use buildbtw::{
 #[tokio::test]
 #[ignore = "Test depends on an external resource and is flaky."]
 async fn test_create_source_repo_cache() -> Result<()> {
-    let source_repo_dir = storage::package_source_repos_dir()?;
+    let source_repo_dir = storage::package_source_repos_dir(&None)?;
     let mut source_repos = dependency_graph::SourceRepoCache::new(&source_repo_dir).await?;
     let mut count = 0;
     for (_dir, repo) in source_repos.all_repos_mut() {
@@ -31,7 +31,7 @@ async fn test_create_source_repo_cache() -> Result<()> {
 #[tokio::test]
 #[ignore = "Test depends on an external resource and is flaky."]
 async fn test_build_buildspace_source_info_index() -> Result<()> {
-    let source_repo_dir = storage::package_source_repos_dir()?;
+    let source_repo_dir = storage::package_source_repos_dir(&None)?;
     let mut source_repos = dependency_graph::SourceRepoCache::new(&source_repo_dir).await?;
     let index = dependency_graph::BuildspaceSourceInfoIndex::build(
         git::Changesets::from(vec![git::Changeset {
@@ -62,7 +62,7 @@ async fn test_build_buildspace_source_info_index() -> Result<()> {
 #[ignore = "Test depends on an external resource and is flaky."]
 async fn test_build_global_dependency_graphs() -> Result<()> {
     // prepare required data
-    let source_repo_dir = storage::package_source_repos_dir()?;
+    let source_repo_dir = storage::package_source_repos_dir(&None)?;
     let mut source_repos = dependency_graph::SourceRepoCache::new(&source_repo_dir).await?;
     let index = dependency_graph::BuildspaceSourceInfoIndex::build(
         git::Changesets::from(vec![git::Changeset {
@@ -104,7 +104,7 @@ async fn test_build_global_dependency_graphs() -> Result<()> {
 #[tokio::test]
 #[ignore = "Test depends on an external resource and is flaky."]
 async fn test_calculate_build_graphs() -> Result<()> {
-    let source_repo_dir = storage::package_source_repos_dir()?;
+    let source_repo_dir = storage::package_source_repos_dir(&None)?;
     let mut source_repos = dependency_graph::SourceRepoCache::new(&source_repo_dir).await?;
 
     // Test creating a build graph for an arbitrary changeset